+++ src/vector/_methods.py
@@ -8,6 +8,7 @@ from __future__ import annotations
 import typing
 from contextlib import suppress
 
+import vector
 from vector._typeutils import (
     BoolCollection,
     ScalarCollection,
@@ -144,7 +145,9 @@ class VectorProtocol:
             vectors without momentum-synonyms.
     """
 
-    lib: Module
+    @property
+    def lib(self) -> Module:
+        ...
 
     def _wrap_result(
         self,
@@ -1348,6 +1351,40 @@ class MomentumProtocolLorentz(VectorProtocolLorentz, MomentumProtocolSpatial):
 
 
 class Vector(VectorProtocol):
+    @typing.overload
+    def __new__(cls, *, x: float, y: float) -> vector.VectorObject2D:
+        ...
+
+    @typing.overload
+    def __new__(cls, *, rho: float, phi: float) -> vector.VectorObject2D:
+        ...
+
+    @typing.overload
+    def __new__(cls, *, px: float, py: float) -> vector.MomentumObject2D:
+        ...
+
+    @typing.overload
+    def __new__(cls, *, x: float, py: float) -> vector.MomentumObject2D:
+        ...
+
+    @typing.overload
+    def __new__(cls, *, px: float, y: float) -> vector.MomentumObject2D:
+        ...
+
+    @typing.overload
+    def __new__(cls, *, pt: float, phi: float) -> vector.MomentumObject2D:
+        ...
+
+    @typing.overload
+    def __new__(cls, __azumthal: Azimuthal) -> Vector:
+        ...
+
+    def __new__(cls, *args: Azimuthal, **kwargs: float) -> Vector:
+        if cls is not Vector:
+            return super().__new__(cls)
+
+        return vector.obj(*args, **kwargs)
+
     def to_xy(self) -> VectorProtocolPlanar:
         from vector._compute import planar
 
@@ -2671,10 +2708,11 @@ def _lib_of(*objects: VectorProtocol) -> Module:  # NumPy-like module
     for obj in objects:
         if isinstance(obj, Vector):
             if lib is None:
-                lib = obj.lib
-            elif lib is not obj.lib:
+                # Not sure why mypy complains about member not being assignable here
+                lib = obj.lib  # type: ignore[misc]
+            elif lib is not obj.lib:  # type: ignore[misc]
                 raise TypeError(
-                    f"cannot use {lib} and {obj.lib} in the same calculation"
+                    f"cannot use {lib} and {obj.lib} in the same calculation"  # type: ignore[misc]
                 )
 
     assert lib is not None