+++ tests/backends/test_object.py
@@ -11,6 +11,66 @@ import pytest
 import vector
 
 
+def test_constructors_2D():
+    vec = vector.VectorObject2D(x=1, y=2)
+    assert vec.x == 1
+    assert vec.y == 2
+
+    vec = vector.VectorObject2D(rho=1, phi=2)
+    assert vec.rho == 1
+    assert vec.phi == 2
+
+    vec = vector.VectorObject2D(
+        azimuthal=vector.backends.object.AzimuthalObjectXY(1, 2)
+    )
+    assert vec.x == 1
+    assert vec.y == 2
+
+    with pytest.raises(TypeError):
+        vector.VectorObject2D(rho=1, wow=2)
+
+    with pytest.raises(TypeError):
+        vector.VectorObject2D(rho=complex(1, 2), wow=2)
+
+    with pytest.raises(TypeError):
+        vector.VectorObject2D()
+
+    vec = vector.MomentumObject2D(px=1, py=2)
+    assert vec.px == 1
+    assert vec.py == 2
+    assert vec.x == 1
+    assert vec.y == 2
+
+    vec = vector.MomentumObject2D(x=1, py=2)
+    assert vec.px == 1
+    assert vec.py == 2
+    assert vec.x == 1
+    assert vec.y == 2
+
+    vec = vector.MomentumObject2D(pt=1, phi=2)
+    assert vec.pt == 1
+    assert vec.phi == 2
+    assert vec.rho == 1
+    assert vec.phi == 2
+
+    vec = vector.MomentumObject2D(
+        azimuthal=vector.backends.object.AzimuthalObjectXY(1, 2)
+    )
+    assert vec.px == 1
+    assert vec.py == 2
+    assert vec.x == 1
+    assert vec.y == 2
+
+    with pytest.raises(TypeError):
+        vector.MomentumObject2D(rho=1, wow=2)
+
+    with pytest.raises(TypeError):
+        vector.MomentumObject2D(rho=False, wow=2)
+
+    with pytest.raises(TypeError):
+        vector.MomentumObject2D()
+
+
 def test_array_casting():
     obj = vector.obj(x=1, y=1)
     assert isinstance(obj, vector.VectorObject2D)