+++ tests/compute/planar/test_deltaphi.py
@@ -14,39 +14,39 @@ import vector.backends.object
 
 def test_xy_xy():
     v1 = vector.backends.object.VectorObject2D(
-        vector.backends.object.AzimuthalObjectXY(1, 2)
+        azimuthal=vector.backends.object.AzimuthalObjectXY(1, 2)
     )
     v2 = vector.backends.object.VectorObject2D(
-        vector.backends.object.AzimuthalObjectXY(3, 4)
+        azimuthal=vector.backends.object.AzimuthalObjectXY(3, 4)
     )
     assert v1.deltaphi(v2) == pytest.approx(math.atan2(2, 1) - math.atan2(4, 3))
 
 
 def test_xy_rhophi():
     v1 = vector.backends.object.VectorObject2D(
-        vector.backends.object.AzimuthalObjectXY(1, 2)
+        azimuthal=vector.backends.object.AzimuthalObjectXY(1, 2)
     )
     v2 = vector.backends.object.VectorObject2D(
-        vector.backends.object.AzimuthalObjectRhoPhi(3, 4)
+        azimuthal=vector.backends.object.AzimuthalObjectRhoPhi(3, 4)
     )
     assert v1.deltaphi(v2) == pytest.approx(math.atan2(2, 1) - 4)
 
 
 def test_rhophi_xy():
     v1 = vector.backends.object.VectorObject2D(
-        vector.backends.object.AzimuthalObjectRhoPhi(1, 2)
+        azimuthal=vector.backends.object.AzimuthalObjectRhoPhi(1, 2)
     )
     v2 = vector.backends.object.VectorObject2D(
-        vector.backends.object.AzimuthalObjectXY(3, 4)
+        azimuthal=vector.backends.object.AzimuthalObjectXY(3, 4)
     )
     assert v1.deltaphi(v2) == pytest.approx(2 - math.atan2(4, 3))
 
 
 def test_rhophi_rhophi():
     v1 = vector.backends.object.VectorObject2D(
-        vector.backends.object.AzimuthalObjectRhoPhi(1, 2)
+        azimuthal=vector.backends.object.AzimuthalObjectRhoPhi(1, 2)
     )
     v2 = vector.backends.object.VectorObject2D(
-        vector.backends.object.AzimuthalObjectRhoPhi(3, 4)
+        azimuthal=vector.backends.object.AzimuthalObjectRhoPhi(3, 4)
     )
     assert v1.deltaphi(v2) == pytest.approx(2 - 4)